HeapSort.py
This is a sorting algorithm that uses a binary heap data structure.  This heap sort is to arrange elements in increasing order. A binary heap is a complete binary tree in which the parent node is greater than the child node or vice-versa. 

Algorithm for increasing order:
First, we arrange items in the max heap.
Second,  the largest element is stored at the top of the heap or say the root of the heap. Now, replace the root item with the last item of the heap then reducing the size of heap by 1. Then heapify the root.
Third, repeat the second step while the size of the heap is greater than 1.